pyproject.toml
[project]
name = "pixverse-mcp"
version = "0.2.2"
description = "Model Context Protocol server for Pixverse video generation APIs"
authors = [
    {name = "Pixverse Team", email = "dev@pixverse.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp[cli]>=1.6.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "asyncio-throttle>=1.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/pixverse/pixverse-mcp"
Repository = "https://github.com/pixverse/pixverse-mcp"
Documentation = "https://pixverse-mcp.readthedocs.io"
Issues = "https://github.com/pixverse/pixverse-mcp/issues"

[project.scripts]
pixverse-mcp = "pixverse_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixverse_mcp"]

[tool.hatch.build.targets.wheel.sources]
"src/pixverse_mcp" = "pixverse_mcp"

[tool.uvx]
entry-point = "pixverse_mcp.__main__:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["pixverse_mcp"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/pixverse_mcp",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[dependency-groups]
dev = [
    "twine>=6.2.0",
]
